Add deploy as default values

+++ app-cli/lib/deploy.js
@@ -6,10 +6,10 @@ const { printError, printMessage } = require("../utils/messageHandler")
 const {
   validateIsGitRepository,
   validateDefinitionFile,
-  optionIsGiven
+  optionIsGiven,
+  validateAuthentication
 } = require("../utils/validationsHandler")
-const { requestLogin } = require("../lib/login")
-const { getFileStream, uploadFile } = require("../utils/fileHelper")
+const { uploadFile } = require("../utils/fileHelper")
 const { gitArchiveFile } = require("../utils/fileHelper")
 const fs = require("fs-extra")
 const path = require("path")
@@ -20,7 +20,7 @@ function initMachineFromLocalStorage() {
   const possibleApp = MachineHelper.apps.find(app => app.cwd === process.cwd())
 
   if (possibleApp) {
-    DeployApi.setMachineToDeeploy(possibleApp.machineToDeploy)
+    DeployApi.setMachineToDeploy(possibleApp.machineToDeploy)
 
     DeployApi.setAppName(possibleApp.appName)
 
@@ -28,30 +28,31 @@ function initMachineFromLocalStorage() {
   }
 }
 
-function deployAsDefaultValues() {
-  const { appName, branchToPush, machineToDeploy } = DeployApi
+async function deployAsDefaultValues() {
+  try {
+    const isValidAuthentication = await validateAuthentication()
 
-  if (!appName || !branchToPush || !machineToDeploy) {
-    printError(
-      "Default deploy failed. There are no default options selected.",
-      true
-    )
-  }
+    if (isValidAuthentication) {
+      // Refresh token in DeployApi
+      initMachineFromLocalStorage()
 
-  printMessage(`Deploying to ${machineToDeploy.name}`)
+      const { appName, branchToPush, machineToDeploy } = DeployApi
 
-  deployFromGitProject()
-}
+      if (!appName || !branchToPush || !machineToDeploy) {
+        printError(
+          "Default deploy failed. There are no default options selected.",
+          true
+        )
+      }
 
-async function validateAuthentication() {
-  // 1. Check if valid auth
-  const isAuthenticated = await DeployApi.isAuthTokenValid()
+      printMessage(`Deploying to ${machineToDeploy.name}`)
 
-  // 2. Request login
-  // 3. Login
-  // 4. Update token
-  if (!isAuthenticated) {
-    requestLogin()
+      deployFromGitProject()
+    } else {
+      printError("Incorrect login details", true)
+    }
+  } catch (e) {
+    printError(e.message, true)
   }
 }
 
@@ -66,28 +67,29 @@ async function deployAsStateless(stateless, host, appName, branch, pass) {
 
     printMessage(`Starting stateless deploy to\n${host}\n${branch}\n${appName}`)
 
-    await deployFromGitProject()
+    deployFromGitProject()
   }
 }
 
 async function deployFromTarFile(tarFile) {
   try {
-    await validateAuthentication()
+    const isValidAuthentication = await validateAuthentication()
 
-    // Send from tar file
-    const filePath = tarFile
-    const fileStream = getFileStream(filePath)
-    const gitHash = "sendviatarfile"
+    if (isValidAuthentication) {
+      // Send from tar file
+      const filePath = tarFile
+      const gitHash = "sendviatarfile"
 
-    await uploadFile(filePath, fileStream, gitHash)
+      await uploadFile(filePath, gitHash)
+    } else {
+      printError("Incorrect login details", true)
+    }
   } catch (e) {
     printError(e.message, true)
   }
 }
 
 function deployFromGitProject() {
-  // TODO - Deploy from git folder
-
   if (!commandExistsSync("git")) {
     printError("'git' command not found...")
 
@@ -104,15 +106,19 @@ function deployFromGitProject() {
 
   // Removes the temporarly file created
   try {
-    fs.removeSync(zipFileFullPath)
+    const tempFileExists = fs.pathExistsSync(zipFileFullPath)
+
+    if (tempFileExists) {
+      fs.removeSync(zipFileFullPath)
+    }
   } catch (e) {
     // IgnoreError
   }
 
-  gitArchiveFile(zipFileFullPath, MachineHelper.branchToPush)
+  gitArchiveFile(zipFileFullPath, DeployApi.branchToPush)
 }
 
-function deploy(options) {
+async function deploy(options) {
   // Reads local storage and sets the machine if found
   initMachineFromLocalStorage()
 
@@ -124,40 +130,44 @@ function deploy(options) {
 
   printMessage("Preparing deployment to Captain...\n")
 
-  const questions = [
-    {
-      type: "list",
-      name: "captainNameToDeploy",
-      default: DeployApi.machineToDeploy.name || "",
-      message: "Select the Captain Machine you want to deploy to:",
-      choices: MachineHelper.getMachinesAsOptions(),
-      when: () => optionIsGiven(options, "host")
-    },
-    {
-      type: "input",
-      default: DeployApi.branchToPush || "master",
-      name: "branchToPush",
-      message: "Enter the 'git' branch you would like to deploy:",
-      when: () => optionIsGiven(options, "branch")
-    },
-    {
-      type: "input",
-      default: DeployApi.appName,
-      name: "appName",
-      message: "Enter the Captain app name this directory will be deployed to:",
-      when: () => optionIsGiven(options, "appName")
-    },
-    {
-      type: "confirm",
-      name: "confirmedToDeploy",
-      message:
-        "Note that uncommitted files and files in gitignore (if any) will not be pushed to server. Please confirm so that deployment process can start.",
-      default: true,
-      when: () => optionIsGiven(options, "stateless")
-    }
-  ]
+  if (options.default) {
+    deployAsDefaultValues()
+  } else {
+    const questions = [
+      {
+        type: "list",
+        name: "captainNameToDeploy",
+        default: DeployApi.machineToDeploy.name || "",
+        message: "Select the Captain Machine you want to deploy to:",
+        choices: MachineHelper.getMachinesAsOptions(),
+        when: () => optionIsGiven(options, "host")
+      },
+      {
+        type: "input",
+        default: DeployApi.branchToPush || "master",
+        name: "branchToPush",
+        message: "Enter the 'git' branch you would like to deploy:",
+        when: () => optionIsGiven(options, "branch")
+      },
+      {
+        type: "input",
+        default: DeployApi.appName,
+        name: "appName",
+        message:
+          "Enter the Captain app name this directory will be deployed to:",
+        when: () => optionIsGiven(options, "appName")
+      },
+      {
+        type: "confirm",
+        name: "confirmedToDeploy",
+        message:
+          "Note that uncommitted files and files in gitignore (if any) will not be pushed to server. Please confirm so that deployment process can start.",
+        default: true,
+        when: () => optionIsGiven(options, "stateless")
+      }
+    ]
+    const answers = await inquirer.prompt(questions)
 
-  inquirer.prompt(questions).then(answers => {
     if (!answers.confirmedToDeploy) {
       printMessage("\nOperation cancelled by the user...\n")
 
@@ -180,16 +190,13 @@ function deploy(options) {
       deployFromTarFile(options.tarFile)
     }
 
-    // if (options.default) {
-    //   deployAsDefaultValues()
-    // }
-
     // if (options.stateless) {
     //   deployAsStateless()
     // }
 
-    // deployFromGitProject()
-  })
+    // Normal deploy
+    deployFromGitProject()
+  }
 }
 
 module.exports = deploy
